Speed improvement and loved tracks support
Now the library is fetched first, so that the API request count is drastically lowered. However, if you don't tag your local collection according to MusicBrainz, you might miss a lot of track matches.

The --updateloved switch now also scans for the loved tracks and updates rating if not already set.